[기능] 스케줄 타입별 목록 조회 피드백 반영

+++ src/main/java/com/damo/server/domain/schedule/ScheduleService.java
@@ -3,79 +3,64 @@ package com.damo.server.domain.schedule;
 import com.damo.server.application.handler.exception.BadRequestException;
 import com.damo.server.application.handler.exception.NotFoundException;
 import com.damo.server.domain.common.pagination.CustomSchedulePage;
+import com.damo.server.domain.common.pagination.param.SchedulePaginationParam;
 import com.damo.server.domain.schedule.dto.RequestScheduleDto;
 import com.damo.server.domain.schedule.dto.ScheduleDto;
 import com.damo.server.domain.schedule.entity.Schedule;
-import com.damo.server.domain.schedule.entity.ScheduleStatus;
-import com.damo.server.domain.schedule.entity.ScheduleTransaction;
+import com.damo.server.domain.user.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 
 @AllArgsConstructor
 @Service
 public class ScheduleService {
     private final ScheduleRepository scheduleRepository;
+    private final UserService userService;
 
     @Transactional
-    public ScheduleDto save(final RequestScheduleDto scheduleDto) {
+    public void save(final RequestScheduleDto scheduleDto) {
         if (scheduleRepository.existsByDateAndEventAndPersonId(scheduleDto.date(), scheduleDto.event(), scheduleDto.personId())) {
             throw new BadRequestException("스케줄 내에서 동일한 기록이 존재");
         }
-        //TODO: return 시 person 출력 유무 협의 예정
-        return ScheduleMapper.toDto(scheduleRepository.save(ScheduleMapper.toEntity(scheduleDto)));
+        scheduleRepository.save(Schedule.from(scheduleDto));
     }
 
-    public ScheduleDto readSchedule(final Long scheduleId) {
-        // TODO: security로 userId 받으면 조회 시 조건문에 userId 추가
-        return scheduleRepository.findOne(scheduleId).orElseThrow(() -> new NotFoundException("조회할 대상을 찾을 수 없음"));
+    public ScheduleAmount readTotalAmounts(final Long userId) {
+        return scheduleRepository.findTotalAmount(userId).orElseThrow(() -> new NotFoundException("조회할 대상을 찾을 수 없음"));
     }
 
-    public CustomSchedulePage readScheduleList(final Pageable pageable, final Long userId, final String transaction) {
-        Page<ScheduleDto> page = scheduleRepository.findAllByUserId(pageable, userId);
+    public ScheduleDto readSchedule(final Long scheduleId, final Long userId) {
+        final Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(() -> new NotFoundException("조회할 대상을 찾을 수 없음"));
+        if (!schedule.getPerson().getUser().getId().equals(userId)) {
+            throw new BadRequestException("다른 사용자의 스케줄");
+        }
+        return ScheduleDto.from(schedule);
+    }
 
-        Integer totalGiving = page.getContent().stream().filter(dto -> dto.getTransaction().getKey().equals("GIVING")).mapToInt(ScheduleDto::getAmount).reduce(0, Integer::sum);
-        Integer totalReceiving = page.getContent().stream().filter(dto -> dto.getTransaction().getKey().equals("RECEIVING")).mapToInt(ScheduleDto::getAmount).reduce(0, Integer::sum);
-        List<ScheduleDto> scheduleDtos = switch (transaction) {
-            case "TOTAL" -> {
-                yield page.getContent();
-            }
-            case "GIVING" -> {
-                yield page.getContent().stream().filter(dto -> dto.getTransaction().getKey().equals("GIVING")).toList();
-            }
-            case "RECEIVING" -> {
-                yield page.getContent().stream().filter(dto -> dto.getTransaction().getKey().equals("RECEIVING")).toList();
-            }
-            default -> {
-                throw new BadRequestException("조회할 거래 종류 입력 오류");
-            }
-        };
-
-        return new CustomSchedulePage(scheduleDtos, pageable, scheduleDtos.size(), totalGiving, totalReceiving);
+    public CustomSchedulePage readScheduleList(final SchedulePaginationParam param, final Long userId) {
+        Page<ScheduleDto> page = scheduleRepository.findAllByUserId(param.toPageable(), userId, param.getStartedAt(), param.getEndedAt(), param.getTransaction());
+        return param.getStartedAt() != null && param.getEndedAt() != null
+               ? new CustomSchedulePage(page)
+               : new CustomSchedulePage(page, scheduleRepository.findTermTotalAmount(userId, param.getStartedAt(), param.getEndedAt()));
     }
 
     @Transactional
-    public ScheduleDto patchScheduleById(final RequestScheduleDto scheduleDto, final Long scheduleId) {
+    public void patchScheduleById(final RequestScheduleDto scheduleDto, final Long scheduleId, final Long userId) {
         final Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(() -> new NotFoundException("수정할 대상을 찾을 수 없음"));
-        // TODO: security로 userId 받으면 같은 유저인지 판단 조건 추가
-
-        schedule.setDate(scheduleDto.date() != null ? scheduleDto.date() : schedule.getDate());
-        schedule.setAmount(scheduleDto.amount() != null ? scheduleDto.amount() : schedule.getAmount());
-        schedule.setMemo(scheduleDto.memo() != null ? scheduleDto.memo() : schedule.getMemo());
-        schedule.setEvent(scheduleDto.event() != null ? scheduleDto.event() : schedule.getEvent());
-        schedule.setStatus(scheduleDto.status() != null ? ScheduleStatus.valueOf(scheduleDto.status()) : schedule.getStatus());
-        schedule.setTransaction(scheduleDto.transaction() != null ? ScheduleTransaction.valueOf(scheduleDto.transaction()) : schedule.getTransaction());
-
-        return ScheduleMapper.toDto(schedule);
+        if (!schedule.getPerson().getUser().getId().equals(userId)) {
+            throw new BadRequestException("다른 사용자의 스케줄");
+        }
+        schedule.changeInfo(scheduleDto);
     }
 
-    public void removeScheduleById(final Long scheduleId) {
-        this.readSchedule(scheduleId);
-        // TODO: security로 userId 받으면 같은 유저인지 판단 조건 추가
+    public void removeScheduleById(final Long scheduleId, final Long userId) {
+        final Schedule schedule = scheduleRepository.findById(scheduleId).orElseThrow(() -> new NotFoundException("조회할 대상을 찾을 수 없음"));
+        if (!schedule.getPerson().getUser().getId().equals(userId)) {
+            throw new BadRequestException("다른 사용자의 스케줄");
+        }
         scheduleRepository.deleteById(scheduleId);
     }
 }
